--- a/src/sum_dirac_dfcoef/utils.py
+++ b/src/sum_dirac_dfcoef/utils.py
@@ -1,7 +1,8 @@
+import pickle
 import re
 import sys
 from pathlib import Path
-from typing import List
+from typing import Any, List
 
 
 def space_separated_parsing(line: str) -> List[str]:
@@ -10,6 +11,18 @@ def space_separated_parsing(line: str) -> List[str]:
 
 def space_separated_parsing_upper(line: str) -> List[str]:
     return [word.upper() for word in line.rstrip("\n").split(" ") if word != ""]
+
+
+def fast_deepcopy_pickle(obj: Any) -> Any:
+    """Fast deep copy using pickle.
+
+    Args:
+        obj (Any): object to be copied
+
+    Returns:
+        Any: copied object
+    """
+    return pickle.loads(pickle.dumps(obj, protocol=pickle.HIGHEST_PROTOCOL))  # noqa: S301
 
 
 def debug_print(message: str) -> None: